Test before_fork behavior
It turns out that we actually don't want barnes to start in the worker, but instead we want it on the "primary" process. This happens to occur by default if you're using Rails as we have a railtie that boots barnes when rails is loaded which executes in the "primary" process.

This test confirms this behavior but doesn't functionally change anything about barnes.